Change tutorial to accomodate `title`
After the introduction of the title element, using the text function to create a title manually is bad practice. I updated the tutorial accordingly.

To make the goal of the tutorial possible, I added an explanation of set show rules. Furthermore, I mentioned the `block` function. Finally, I changed the document title to title case.